Hopeless refactoring
