+++ app/src/main/java/dev/msfjarvis/aps/util/activity/OpenDocumentTreeWithPermissions.kt
@@ -1,0 +1,27 @@
+/*
+ * Copyright © 2014-2020 The Android Password Store Authors. All Rights Reserved.
+ * SPDX-License-Identifier: GPL-3.0-only
+ */
+
+package dev.msfjarvis.aps.util.activity
+
+import android.content.Context
+import android.content.Intent
+import android.net.Uri
+import androidx.activity.result.contract.ActivityResultContracts
+
+/**
+ * Implementation of [ActivityResultContracts.OpenDocumentTree] that grants additional
+ * read-write permissions to the Uri, allowing for durable access over time.
+ */
+class OpenDocumentTreeWithPermissions : ActivityResultContracts.OpenDocumentTree() {
+
+    override fun createIntent(context: Context, input: Uri?): Intent {
+        return super.createIntent(context, input).apply {
+            flags = Intent.FLAG_GRANT_READ_URI_PERMISSION or
+                Intent.FLAG_GRANT_WRITE_URI_PERMISSION or
+                Intent.FLAG_GRANT_PERSISTABLE_URI_PERMISSION or
+                Intent.FLAG_GRANT_PREFIX_URI_PERMISSION
+        }
+    }
+}
